Consolidate Cytation backends: remove duplicates, port cytation.py logic to aravis driver
- Remove cytation_aravis.py (duplicate of cytation_aravis_driver.py)
- Remove test_aravis_connection.ipynb (referenced deleted module)
- Port set_imaging_mode version 2 commands from cytation.py to aravis driver
- Port capture orchestration (image_size, tiling, timing) to microscopy backend
- Port acquire_image retry with stop/start recovery to aravis driver
- Add SetupParams to CytationAravisDriver (matches cytation.py pattern)
- Add CaptureParams.overlap to microscopy backend
- Rename CytationAravisMicroscopyBackend -> CytationMicroscopyBackend
- Rename cytation_aravis_microscopy.py -> cytation_microscopy_backend.py
